Impressions! Proper linking of Query Id!  Handling of missing Brand!
Too much happened too quick in this.  We now have impression tracking.  Also, had a bug that the customQuery would fire BEFORE onValueChange, which meant that we would use the PREVIOUS search everytime.  onKeyPress happens before customQuery, so we get the id, and the onValueChange happens after custom query, so we can travk the event.